React the test base refactoring, and address build failure.
