+++ src/Http/Http.Extensions/test/RequestDelegateGenerator/Baselines/MapAction_ImplicitQuery_NullableStringArrayParam_StringReturn.generated.txt
@@ -223,7 +223,7 @@ namespace Microsoft.AspNetCore.Http.Generated
             return httpContext.Response.WriteAsJsonAsync(value!, options.GetTypeInfo(runtimeType));
         }
 
-        private static async ValueTask<(bool, T?)> TryResolveBody<T>(HttpContext httpContext, bool allowEmpty)
+        private static async ValueTask<(bool, T?)> TryResolveBodyAsync<T>(HttpContext httpContext, bool allowEmpty)
         {
             var feature = httpContext.Features.Get<Microsoft.AspNetCore.Http.Features.IHttpRequestBodyDetectionFeature>();
 
@@ -254,6 +254,11 @@ namespace Microsoft.AspNetCore.Http.Generated
                     return (false, default);
                 }
             }
+            else if (!allowEmpty)
+            {
+                httpContext.Response.StatusCode = StatusCodes.Status400BadRequest;
+            }
+
             return (false, default);
         }
 
@@ -261,6 +266,18 @@ namespace Microsoft.AspNetCore.Http.Generated
             where T : class, IBindableFromHttpContext<T>
         {
             return T.BindAsync(context, parameter);
+        }
+
+        private static ValueTask<(bool, T?)> TryResolveJsonBodyOrServiceAsync<T>(HttpContext httpContext, bool isOptional, IServiceProviderIsService? serviceProviderIsService = null)
+        {
+            if (serviceProviderIsService is not null)
+            {
+                if (serviceProviderIsService.IsService(typeof(T)))
+                {
+                    return new ValueTask<(bool, T?)>((true, httpContext.RequestServices.GetService<T>()));
+                }
+            }
+            return TryResolveBodyAsync<T>(httpContext, isOptional);
         }
     }
